requirements.txt
backcall==0.1.0
certifi==2019.3.9
chardet==3.0.4
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.1
django-cors-headers==3.0.1
django-extensions==2.1.7
djangorestframework==3.9.4
gunicorn==19.9.0
idna==2.8
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
oauthlib==3.0.1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
pipenv==2018.11.26
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.1
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
sqlparse==0.3.0
traitlets==4.3.2
urllib3==1.25.3
virtualenv==16.6.0
virtualenv-clone==0.5.3
wcwidth==0.1.7
Werkzeug==0.15.4
